requirements.txt
# Core async HTTP
aiohttp>=3.9.0

# Environment variables
python-dotenv>=1.0.0

# Data analysis
pandas>=2.0.0
numpy>=1.24.0

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0

# Development/Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
